+++ activerecord/CHANGELOG.md
@@ -1,4 +1,4 @@
-*   Add `:prefix` option to `enum` definition.
+*   Add `:enum_prefix`/`:enum_suffix` option to `enum` definition.
 
     Fixes #17511 and #17415